feat: add example for WithConsoleTemplateAndTheme method
